Add guide to implementing CS for GitHub orgs
The existing Community Specification templates and "Getting Started" guide generally assume that a single repository will be used to house the governance documentation as well as a single specification.

For projects that will develop multiple specifications or otherwise have more than one repo, it may be useful to have the CS documentation in a "governance" repo, with the specs developed in separate repos. If doing so, there are some conforming changes that should be made to the contents of the CS documents.

The guide committed here walks through the practical steps that a project could use to implement this. The intended audience is a user who is not especially familiar with Git / GitHub and who will be using the GitHub user interface to fork and set up the repos.

Signed-off-by: Steve Winslow <steve@swinslow.net>